Rename SidebarLink to IconLinkItem
