+++ apps/frontend/package.json
